requirements.txt
wheel
PyNaCl
demucs
ffmpeg
asyncio
discord
requests
edge-tts
setuptools
torchcrepe
discord.py
onnxruntime
tensorboard
configparser
tensorboardX
scipy==1.9.3
httpx==0.23.0
numba==0.56.4
numpy==1.23.5
librosa==0.9.1
fairseq==0.12.2
faiss-cpu==1.7.3
pyworld>=0.3.2
soundfile>=0.12.1
praat-parselmouth>=0.4.2